refactor(pack): refactor xdsl router
