--- packages/manager/apps/telecom/src/app/telecom/pack/xdsl/modem/router/pack-xdsl-modem-router.controller.js
@@ -1,7 +1,0 @@
-angular
-  .module('managerApp')
-  .controller('XdslModemRouterCtrl', function XdslModemRouterCtrl(
-    TucPackXdslModemMediator,
-  ) {
-    this.mediator = TucPackXdslModemMediator;
-  });
