+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/modem/router/router.controller.js
@@ -1,0 +1,5 @@
+export default /* @ngInject */ function XdslModemRouterCtrl(
+  TucPackXdslModemMediator,
+) {
+  this.mediator = TucPackXdslModemMediator;
+}
